Request Changes for PR
Fixed File and Directory structure, for cleaner code
