+++ ui/src/components/Navbar/index.tsx
@@ -1,2 +1,2 @@
-import Navbar from "./Navbar";
+import { Navbar } from "./Navbar";
 export default Navbar;
