+++ ui/src/components/Preview/Overview/MetricContent/MetricContent.module.css
@@ -1,0 +1,7 @@
+.wrapper {
+  width: 350px;
+}
+
+.relative {
+  position: relative;
+}
