+++ ui/src/components/Preview/Overview/MetricContent/index.tsx
@@ -1,0 +1,2 @@
+import { MetricContent } from "./MetricContent";
+export default MetricContent;
