+++ ui/src/components/Preview/Overview/MetricInfoIcon/MetricInfoIcon.module.css
@@ -1,0 +1,10 @@
+.metricInfo {
+  position: absolute;
+  top: 10px;
+  right: 30px;
+  z-index: 10;
+}
+
+.infoIcon {
+  opacity: 0.5;
+}
