+++ ui/src/components/Preview/Overview/MetricInfoIcon/index.tsx
@@ -1,0 +1,2 @@
+import { MetricInfoIcon } from "./MetricInfoIcon";
+export default MetricInfoIcon;
