+++ ui/src/components/Preview/Overview/MetricVerticalTabs/VerticalTabs.module.css
@@ -1,0 +1,3 @@
+.verticalTabs {
+  display: flex;
+}
