+++ ui/src/components/Preview/Overview/MetricVerticalTabs/index.tsx
@@ -1,0 +1,2 @@
+import { MetricVerticalTabs } from "./VerticalTabs";
+export default MetricVerticalTabs;
