+++ ui/src/components/Preview/Overview/index.tsx
@@ -1,0 +1,2 @@
+import { Overview } from "./Overview";
+export default Overview;
