+++ ui/src/components/Preview/Snapshot/index.tsx
@@ -1,0 +1,2 @@
+import { Snapshot } from "./Snapshot";
+export default Snapshot;
