--- ui/src/components/Tabs/index.tsx
@@ -1,2 +1,0 @@
-import Tabs from "./Tabs";
-export default Tabs;
