+++ ui/src/components/UI/Piechart/index.tsx
@@ -1,0 +1,2 @@
+import { PieChart } from "./Piechart";
+export default PieChart;
