+++ ui/src/components/UI/Tabs/Tabs.module.css
@@ -1,0 +1,9 @@
+.tabs {
+  display: flex;
+  justify-content: space-between;
+  width: fit-content;
+  margin: 1rem auto;
+  border-radius: 20px;
+  justify-content: center;
+  align-items: center;
+}
